20240320_IT#0829_Excel_データベース_取り込み/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openpyxl
pyodbc
azure-storage-blob